requirements/local.txt
asgiref==3.8.1
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-model-utils==4.4.0
pillow==10.2.0
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2024.1
Unipath==1.1
